Refined the Describe Data, Questions, Wrangling sections
For Describe the data, I made the format as desired in tables and I added more description to each of them. For Questions, I added the steps I need to take to wrangle the data, I also make it more visually appealing. For the Wrangling section, I changed start_time and end time to datetime formats and I decided not to split it into 2. I added observations that I noticed in the plots